feat: finished challenges
completed challenges DayOfTheWeek: using a switch statement display
the day of the week based on a number passed from 0 - 6.

NumberIsPrime: display if a number is a prime number or not
based on a range of numbers generated with a for loop.

SumThreeAndFive: sum up numbers that are divisable between three and five
from 1 to 1000 and break the loop once you have found 5 numbers that
meet the condition.